wallpaperapp/backend/authentication/pyproject.toml
[tool.poetry]
name = "authentication"
version = "0.1.0"
description = ""
authors = ["skorodenko <72010957+skorodenko@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
celery = "^5.3.1"
django-cors-headers = "^4.2.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
django-celery-results = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
